refactor(agentic): split parsers and servers by component
