+++ .gitignore
@@ -48,10 +48,6 @@ span.log
 
 # Data and models (large / local-only)
 data/
-# ...but keep the hand-written VizDoom doc list under version control
-!lmms_eval/tasks/vizdoom_agentic/data/
-lmms_eval/tasks/vizdoom_agentic/data/*
-!lmms_eval/tasks/vizdoom_agentic/data/vizdoom.jsonl
 # ViZDoom runtime config droppings
 _vizdoom.ini
 pretrained/
