+++ lmms_eval/agentic/components.py
@@ -20,16 +20,16 @@ from typing import Any
 
 REGISTRY: dict[str, dict[str, str]] = {
     "model_server": {
-        "openai": "lmms_eval.agentic.servers:OpenAIModelServer",
-        "debug": "lmms_eval.agentic.servers:FixedActionModelServer",
+        "openai": "lmms_eval.agentic.servers.openai:OpenAIModelServer",
+        "debug": "lmms_eval.agentic.servers.debug:FixedActionModelServer",
     },
     "model_output_parser": {
-        "identity": "lmms_eval.agentic.parsers:IdentityModelOutputParser",
-        "qwen": "lmms_eval.agentic.parsers:QwenModelOutputParser",
+        "identity": "lmms_eval.agentic.parsers.model_output:IdentityModelOutputParser",
+        "qwen": "lmms_eval.agentic.parsers.model_output:QwenModelOutputParser",
     },
     "observation_parser": {},
     "action_parser": {
-        "action_name": "lmms_eval.agentic.parsers:ActionNameParser",
+        "action_name": "lmms_eval.agentic.parsers.actions:ActionNameParser",
     },
 }
