+++ lmms_eval/agentic/parsers/__init__.py
@@ -1,0 +1,27 @@
+"""Built-in parser interfaces and implementations.
+
+Classes are re-exported here to preserve the original public import surface
+while keeping each parser family in a focused module.
+"""
+
+from lmms_eval.agentic.parsers.actions import ActionNameParser
+from lmms_eval.agentic.parsers.base import (
+    ActionParser,
+    ModelOutputParser,
+    ObservationParser,
+    ParserContext,
+)
+from lmms_eval.agentic.parsers.model_output import (
+    IdentityModelOutputParser,
+    QwenModelOutputParser,
+)
+
+__all__ = [
+    "ActionNameParser",
+    "ActionParser",
+    "IdentityModelOutputParser",
+    "ModelOutputParser",
+    "ObservationParser",
+    "ParserContext",
+    "QwenModelOutputParser",
+]
